Factor out Project Initiailisation
