+++ inlang/source-code/paraglide/paraglide-js-adapter-next/src/cli/utils.ts
